Move django apps to boards_of_django directory
