+++ boards_of_django/authentication/tests/apis/test_user_confirm_registration.py
