+++ boards_of_django/common/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CommonConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "boards_of_django.common"
