+++ config/urls.py
@@ -35,6 +35,6 @@ urlpatterns = [
     re_path(r"^swagger/$", schema_view.with_ui("swagger", cache_timeout=0), name="schema-swagger-ui"),
     re_path(r"^redoc/$", schema_view.with_ui("redoc", cache_timeout=0), name="schema-redoc"),
     path("admin/", admin.site.urls),
-    path("auth/", include(("authentication.urls", "authentication"))),
-    path("", include(("boards.urls", "boards"))),
+    path("auth/", include(("boards_of_django.authentication.urls", "authentication"))),
+    path("", include(("boards_of_django.boards.urls", "boards"))),
 ]
